whisper/api_gateway/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
grpcio==1.60.0
grpcio-tools==1.60.0
python-multipart==0.0.6
nats-py==2.6.0
sse-starlette==1.8.2
httpx==0.25.2
